+++ pygdbmi/gdbmiparser.py
@@ -137,35 +137,35 @@ def assert_match(actual_char_or_str, expected_char_or_str):
 # In addition to a number of out-of-band notifications,
 # the response to a gdb/mi command includes one of the following result indications:
 # done, running, connected, error, exit
-_GDB_MI_RESULT_RE = re.compile("^(\d*)\^(\S+?)(,(.*))?$")
+_GDB_MI_RESULT_RE = re.compile(r"^(\d*)\^(\S+?)(,(.*))?$")
 
 # https://sourceware.org/gdb/onlinedocs/gdb/GDB_002fMI-Async-Records.html#GDB_002fMI-Async-Records
 # Async records are used to notify the gdb/mi client of additional
 # changes that have occurred. Those changes can either be a consequence
 # of gdb/mi commands (e.g., a breakpoint modified) or a result of target activity
 # (e.g., target stopped).
-_GDB_MI_NOTIFY_RE = re.compile("^(\d*)[*=](\S+?),(.*)$")
+_GDB_MI_NOTIFY_RE = re.compile(r"^(\d*)[*=](\S+?),(.*)$")
 
 # https://sourceware.org/gdb/onlinedocs/gdb/GDB_002fMI-Stream-Records.html#GDB_002fMI-Stream-Records
 # "~" string-output
 # The console output stream contains text that should be displayed
 # in the CLI console window. It contains the textual responses to CLI commands.
-_GDB_MI_CONSOLE_RE = re.compile('~"(.*)"', re.DOTALL)
+_GDB_MI_CONSOLE_RE = re.compile(r'~"(.*)"', re.DOTALL)
 
 # https://sourceware.org/gdb/onlinedocs/gdb/GDB_002fMI-Stream-Records.html#GDB_002fMI-Stream-Records
 # "&" string-output
 # The log stream contains debugging messages being produced by gdb's internals.
-_GDB_MI_LOG_RE = re.compile('&"(.*)"', re.DOTALL)
+_GDB_MI_LOG_RE = re.compile(r'&"(.*)"', re.DOTALL)
 
 # https://sourceware.org/gdb/onlinedocs/gdb/GDB_002fMI-Stream-Records.html#GDB_002fMI-Stream-Records
 # "@" string-output
 # The target output stream contains any textual output from the
 # running target. This is only present when GDB's event loop is truly asynchronous,
 # which is currently only the case for remote targets.
-_GDB_MI_TARGET_OUTPUT_RE = re.compile('@"(.*)"', re.DOTALL)
+_GDB_MI_TARGET_OUTPUT_RE = re.compile(r'@"(.*)"', re.DOTALL)
 
 # Response finished
-_GDB_MI_RESPONSE_FINISHED_RE = re.compile("^\(gdb\)\s*$")
+_GDB_MI_RESPONSE_FINISHED_RE = re.compile(r"^\(gdb\)\s*$")
 
 _WHITESPACE = [" ", "\t", "\r", "\n"]