+++ tests/__main__.py
@@ -1,0 +1,3 @@
+from . import test_pygdbmi, static_tests
+
+exit(test_pygdbmi.main() + static_tests.main())